# Perf: Move multiline xml text encoding out of domToText