--- a/core/field_multilineinput.js
+++ b/core/field_multilineinput.js
@@ -70,6 +70,32 @@ Blockly.FieldMultilineInput.fromJson = function(options) {
 };
 
 /**
+ * Serializes this field's value to XML. Should only be called by Blockly.Xml.
+ * @param {!Element} fieldElement The element to populate with info about the
+ *    field's state.
+ * @return {!Element} The element containing info about the field's state.
+ * @package
+ */
+Blockly.FieldMultilineInput.prototype.toXml = function(fieldElement) {
+  // Replace '\n' characters with html-escaped equivalent '&#10'. This is needed
+  // so the plain-text representation of the xml produced by `Blockly.Xml.domToText`
+  // will appear on a single line (this is a limitation of the plain-text format).
+  fieldElement.textContent = this.getValue().replace(/\n/g, '&#10;');
+  return fieldElement;
+};
+
+/**
+ * Sets the field's value based on the given XML element. Should only be
+ * called by Blockly.Xml.
+ * @param {!Element} fieldElement The element containing info about the
+ *    field's state.
+ * @package
+ */
+Blockly.FieldMultilineInput.prototype.fromXml = function(fieldElement) {
+  this.setValue(fieldElement.textContent.replace(/&#10;/g, '\n'));
+};
+
+/**
  * Create the block UI for this field.
  * @package
  */