--- a/core/field_multilineinput.js
+++ b/core/field_multilineinput.js
@@ -77,9 +77,10 @@ Blockly.FieldMultilineInput.fromJson = function(options) {
  * @package
  */
 Blockly.FieldMultilineInput.prototype.toXml = function(fieldElement) {
-  // Replace '\n' characters with html-escaped equivalent '&#10'. This is needed
-  // so the plain-text representation of the xml produced by `Blockly.Xml.domToText`
-  // will appear on a single line (this is a limitation of the plain-text format).
+  // Replace '\n' characters with html-escaped equivalent '&#10'. This is
+  // needed so the plain-text representation of the xml produced by
+  // `Blockly.Xml.domToText` will appear on a single line (this is a limitation
+  // of the plain-text format).
   fieldElement.textContent = this.getValue().replace(/\n/g, '&#10;');
   return fieldElement;
 };