## The basics

- [x] I branched from develop
- [x] My pull request is against develop
- [x] My code follows the [style guide](https://developers.google.com/blockly/guides/modify/web/style-guide)

## The details

A complex regular expression in `Blockly.Xml.domToText` was running on every save, causing performance problems on large projects. The regexp progressively replaces all `\n` with `&#10;` in content fields, and was added along with `FieldMultilineInput`, in #2663.

### Proposed Changes

Override `toXml` and `fromXml` on `FieldMultilineInput` and do the conversion there. This simplifies the regexp, and it only runs when there is multiline content in the program, improving performance - especially on large programs.

### Performance Comparison

Testing with this MakeCode Arcade game: https://arcade.makecode.com/74548-67333-76127-33508

Time spent in `Blockly.Xml.domToText`:

| Before | After |
| -- | -- |
| 10s | 250ms |



### Resolves

Resolves #4030

### Test Coverage

I tested this change in playground.html. Steps I followed:
1. Add a few paragraph blocks and populate them with multiline content.
2. Export to Xml. Inspect the Xml to verify content encoding.
3. Delete all blocks.
4. Import from Xml. Verify (visually) that the content looks correct.

Also: In the dev console, set breakpoints in the `toXml` and `fromXml` methods of `FieldMultilineInput` to verify they're getting called at the right times.

Tested on:
* Desktop Chrome

### Additional information

**Of note:** Another document-wide regexp has appeared in `domToText`, and will likely become a performance issue. You may want to fix it: https://github.com/google/blockly/blob/ba68081d8f0f8e81d76baf370716d5955bae82a1/core/xml.js#L327

I can see how `domToText` would be a tempting place to fixup formatting issues. It might be worth adding a comment in that method stating: "For performance reasons, don't run regular expressions on the whole document."