@eanders-MS 2020-07-14
@googlebot I signed it!


From: googlebot <notifications@github.com>
Sent: Tuesday, July 14, 2020 12:49 PM
To: google/blockly <blockly@noreply.github.com>
Cc: Eric Anderson <eanders@microsoft.com>; Author <author@noreply.github.com>
Subject: Re: [google/blockly] Perf: Move multiline xml text encoding out of domToText (#4046)


Thanks for your pull request. It looks like this may be your first contribution to a Google open source project (if not, look below for help). Before we can look at your pull request, you'll need to sign a Contributor License Agreement (CLA).

📝 Please visit https://cla.developers.google.com/<https://nam06.safelinks.protection.outlook.com/?url=https%3A%2F%2Fcla.developers.google.com%2F&data=02%7C01%7Ceanders%40microsoft.com%7C5a55d9f133d34df5150f08d8282ee874%7C72f988bf86f141af91ab2d7cd011db47%7C1%7C0%7C637303529214799915&sdata=5T58krPOjQdAlEW7QpFnivJT3BlcFc54e9a5e5VHtyE%3D&reserved=0> to sign.

Once you've signed (or fixed any issues), please reply here with @googlebot I signed it! and we'll verify it.

________________________________
What to do if you already signed the CLA
Individual signers

  *   It's possible we don't have your GitHub username or you're using a different email address on your commit. Check your existing CLA data<https://nam06.safelinks.protection.outlook.com/?url=https%3A%2F%2Fcla.developers.google.com%2Fclas&data=02%7C01%7Ceanders%40microsoft.com%7C5a55d9f133d34df5150f08d8282ee874%7C72f988bf86f141af91ab2d7cd011db47%7C1%7C0%7C637303529214809909&sdata=Wnn8yvNzb12z29ioZTCyONwicLB3tTg6NhCZVCmdUHo%3D&reserved=0> and verify that your email is set on your git commits<https://nam06.safelinks.protection.outlook.com/?url=https%3A%2F%2Fhelp.github.com%2Farticles%2Fsetting-your-email-in-git%2F&data=02%7C01%7Ceanders%40microsoft.com%7C5a55d9f133d34df5150f08d8282ee874%7C72f988bf86f141af91ab2d7cd011db47%7C1%7C0%7C637303529214809909&sdata=E%2FxkZgxl7UBtymMD2iyW9zYC6I8gyyrkimiF0TTEkvQ%3D&reserved=0>.

Corporate signers

  *   Your company has a Point of Contact who decides which employees are authorized to participate. Ask your POC to be added to the group of authorized contributors. If you don't know who your Point of Contact is, direct the Google project maintainer to go/cla#troubleshoot<http://go/cla#troubleshoot> (Public version<https://nam06.safelinks.protection.outlook.com/?url=https%3A%2F%2Fopensource.google%2Fdocs%2Fcla%2F%23troubleshoot&data=02%7C01%7Ceanders%40microsoft.com%7C5a55d9f133d34df5150f08d8282ee874%7C72f988bf86f141af91ab2d7cd011db47%7C1%7C0%7C637303529214819899&sdata=FRTB%2B3cOjM01lAHae0CAEFhbbsAfPLvhzeadaQv%2F8BA%3D&reserved=0>).
  *   The email used to register you as an authorized contributor must be the email used for the Git commit. Check your existing CLA data<https://nam06.safelinks.protection.outlook.com/?url=https%3A%2F%2Fcla.developers.google.com%2Fclas&data=02%7C01%7Ceanders%40microsoft.com%7C5a55d9f133d34df5150f08d8282ee874%7C72f988bf86f141af91ab2d7cd011db47%7C1%7C0%7C637303529214819899&sdata=0sSXMNlaBqwo%2FG5VgKWcG3XeVfJHv4Q0ugq%2F0dlqnzQ%3D&reserved=0> and verify that your email is set on your git commits<https://nam06.safelinks.protection.outlook.com/?url=https%3A%2F%2Fhelp.github.com%2Farticles%2Fsetting-your-email-in-git%2F&data=02%7C01%7Ceanders%40microsoft.com%7C5a55d9f133d34df5150f08d8282ee874%7C72f988bf86f141af91ab2d7cd011db47%7C1%7C0%7C637303529214829894&sdata=I%2BU6ucaFGxDjOEy9J5bozFSFLkbZG4NUFqN9O71K4vI%3D&reserved=0>.
  *   The email used to register you as an authorized contributor must also be attached to your GitHub account<https://nam06.safelinks.protection.outlook.com/?url=https%3A%2F%2Fgithub.com%2Fsettings%2Femails&data=02%7C01%7Ceanders%40microsoft.com%7C5a55d9f133d34df5150f08d8282ee874%7C72f988bf86f141af91ab2d7cd011db47%7C1%7C0%7C637303529214829894&sdata=MdhAuZl6e7udaQt4eBEYXX2JQiNZlA6aBSZuEaS5rYw%3D&reserved=0>.

ℹ️ Googlers: Go here<https://nam06.safelinks.protection.outlook.com/?url=https%3A%2F%2Fgoto.google.com%2Fprinfo%2Fhttps%253A%252F%252Fgithub.com%252Fgoogle%252Fblockly%252Fpull%252F4046&data=02%7C01%7Ceanders%40microsoft.com%7C5a55d9f133d34df5150f08d8282ee874%7C72f988bf86f141af91ab2d7cd011db47%7C1%7C0%7C637303529214839893&sdata=1K60CseeNDbA6JTXtT3GrqdUP58aGXLzOBWRiwd36Cg%3D&reserved=0> for more info.

—
You are receiving this because you authored the thread.
Reply to this email directly, view it on GitHub<https://nam06.safelinks.protection.outlook.com/?url=https%3A%2F%2Fgithub.com%2Fgoogle%2Fblockly%2Fpull%2F4046%23issuecomment-658378346&data=02%7C01%7Ceanders%40microsoft.com%7C5a55d9f133d34df5150f08d8282ee874%7C72f988bf86f141af91ab2d7cd011db47%7C1%7C0%7C637303529214839893&sdata=oySxxacQ6YMockJWVfohIdYhnUdVQj8eRHHCSc%2B1mok%3D&reserved=0>, or unsubscribe<https://nam06.safelinks.protection.outlook.com/?url=https%3A%2F%2Fgithub.com%2Fnotifications%2Funsubscribe-auth%2FAC443J4XOT266TGUV74CAHTR3SZBPANCNFSM4OZ23U4Q&data=02%7C01%7Ceanders%40microsoft.com%7C5a55d9f133d34df5150f08d8282ee874%7C72f988bf86f141af91ab2d7cd011db47%7C1%7C0%7C637303529214849880&sdata=pST5u972Itr9A8HGqiHcUetKqxwocAMLxh2nK%2BbTb90%3D&reserved=0>.